requirements.txt
# BUNK3R_IA - Consolidated Dependencies v2.0

# Core Flask & Web
flask>=2.3.0
gunicorn>=21.0.0
flask-cors>=4.0.0
flask-sqlalchemy
flask-login
flask-dance
psycopg2-binary>=2.9.0

# API / Fast API
fastapi
uvicorn
pydantic

# AI Providers & HTTP
requests>=2.31.0
httpx>=0.25.0
python-dotenv>=1.0.0
pyjwt

# Automation & Tools
playwright>=1.40.0
pyautogui
pyperclip

# Testing
pytest
websocket-client
